feat(26075204): refactor code
