+++ src/main.tsx
@@ -1,11 +1,11 @@
 import { createRoot } from 'react-dom/client'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import Home from './pages/index'
+import Home from '@/pages/index'
+import Login from '@/pages/login'
+import Email from '@/pages/email'
+import LockScreen from '@/pages/lock-screen'
+import ResetPassword from '@/pages/reset-password'
 import './style.css'
-import Login from './pages/login'
-import Email from './pages/email'
-import LockScreen from './pages/lockscreen'
-import Reset from './pages/reset'
 
 const App = (): JSX.Element => {
   const router = createBrowserRouter([
@@ -24,13 +24,13 @@ const App = (): JSX.Element => {
     },
 
     {
-      path: '/lockScreen',
+      path: '/lock-screen',
       element: <LockScreen />
     },
 
     {
-      path: '/resetPassword',
-      element: <Reset />
+      path: '/reset-password',
+      element: <ResetPassword />
     }
   ])
