+++ src/pages/index.tsx
@@ -1,7 +1,6 @@
-import Discover from '../components/Discover'
-import Header from '../components/Header'
-import Title from '../components/Title'
-import Footer from '../components/Footer'
+import Discover from '@/components/Discover'
+import Footer from '@/components/Footer'
+import Title from '@/components/Title'
 
 const Home = (): JSX.Element => {
   return (
